+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - make the TCA cacheable
 
 ### Changed
+- #1 Convert the locallang files to XLIFF
 - move the extension to GitHub
 - allow oelib up to 1.9.99
 - set 7.9.99 as maximum TYPO3 version